refactoring the code - preventing circular import error, moving out common code to different file
