+++ dev/breeze/src/airflow_breeze/cache.py
@@ -20,10 +20,8 @@ from pathlib import Path
 from typing import Any, List, Optional, Tuple
 
 from airflow_breeze import global_constants
-from airflow_breeze.breeze import get_airflow_sources_root
 from airflow_breeze.console import console
-
-BUILD_CACHE_DIR = Path(get_airflow_sources_root(), '.build')
+from airflow_breeze.global_constants import BUILD_CACHE_DIR
 
 
 def check_if_cache_exists(param_name: str) -> bool:
@@ -43,6 +41,7 @@ def write_to_cache_file(param_name: str, param_value: str, check_allowed_values:
     if check_allowed_values:
         allowed, allowed_values = check_if_values_allowed(param_name, param_value)
     if allowed or not check_allowed_values:
+        print('BUID CACHE DIR:', BUILD_CACHE_DIR)
         Path(BUILD_CACHE_DIR, f".{param_name}").write_text(param_value)
     else:
         console.print(f'[cyan]You have sent the {param_value} for {param_name}')
