+++ dev/breeze/src/airflow_breeze/global_constants.py
@@ -14,15 +14,19 @@
 # KIND, either express or implied.  See the License for the
 # specific language governing permissions and limitations
 # under the License.
+import os
+from pathlib import Path
+from typing import Optional
+
+from airflow_breeze.console import console
 
 AIRFLOW_SOURCES = ""
 DEFAULT_PYTHON_MAJOR_MINOR_VERSION = 3.7
-BUILD_CACHE_DIR = "${AIRFLOW_SOURCES}/.build"
 
 FORCE_PULL_IMAGES = False
 CHECK_IF_BASE_PYTHON_IMAGE_UPDATED = False
 FORCE_BUILD_IMAGES = False
-LAST_FORCE_ANSWER_FILE = f"{BUILD_CACHE_DIR}/last_force_answer.sh"
+# LAST_FORCE_ANSWER_FILE = f"{BUILD_CACHE_DIR}/last_force_answer.sh"
 FORCE_ANSWER_TO_QUESTION = ""
 SKIP_CHECK_REMOTE_IMAGE = False
 PUSH_PYTHON_BASE_IMAGE = False
@@ -177,3 +181,41 @@ PARAM_NAME_FLAG = {
     "POSTGRES_VERSION": "--postgres-version",
     "MSSQL_VERSION": "--mssql-version",
 }
+
+
+__AIRFLOW_SOURCES_ROOT = Path.cwd()
+
+__AIRFLOW_CFG_FILE = "setup.cfg"
+
+
+def get_airflow_sources_root():
+    return __AIRFLOW_SOURCES_ROOT
+
+
+def search_upwards_for_airflow_sources_root(start_from: Path) -> Optional[Path]:
+    root = Path(start_from.root)
+    d = start_from
+    while d != root:
+        attempt = d / __AIRFLOW_CFG_FILE
+        if attempt.exists() and "name = apache-airflow\n" in attempt.read_text():
+            return attempt.parent
+        d = d.parent
+    return None
+
+
+def find_airflow_sources_root():
+    # Try to find airflow sources in current working dir
+    airflow_sources_root = search_upwards_for_airflow_sources_root(Path.cwd())
+    if not airflow_sources_root:
+        # Or if it fails, find it in parents of the directory where the ./breeze.py is.
+        airflow_sources_root = search_upwards_for_airflow_sources_root(Path(__file__).resolve().parent)
+    global __AIRFLOW_SOURCES_ROOT
+    if airflow_sources_root:
+        __AIRFLOW_SOURCES_ROOT = airflow_sources_root
+    else:
+        console.print(f"\n[yellow]Could not find Airflow sources location. Assuming {__AIRFLOW_SOURCES_ROOT}")
+    os.chdir(__AIRFLOW_SOURCES_ROOT)
+
+
+find_airflow_sources_root()
+BUILD_CACHE_DIR = Path(get_airflow_sources_root(), '.build')
